+++ routes/api/cars.js
@@ -1,67 +1,64 @@
+require("dotenv").config();
 const express = require('express');
 const router = express.Router();
-const path = require('path');
 const config = require('../../config.json')
 const carsDB = require('../../models/schemas/cars')
-const { v4: uuidv4 } = require('uuid')
-const { authCheck } = require('../../middleware/authentication/authentication');
-const { error } = require('console');
+const { apiAuthCheck } = require('../../middleware/authentication/apiAuthentication');
+const { checkOrigin } = require('../../middleware/authentication/checkOrigin');
 
+const AWS = require('aws-sdk')
+const multer = require('multer');
+
+// Configure AWS SDK
+AWS.config.update({
+    accessKeyId: process.env.AWS_ACCESS_KEY,
+    secretAccessKey: process.env.AWS_SECRET_KEY,
+    region: process.env.AWS_REGION
+});
+
+const s3 = new AWS.S3();
+
+// Multer upload middleware
+const upload = multer({
+    limits: {
+        fileSize: 100 * 1024 * 1024, // 100MB, adjust as needed
+        files: 10 // Maximum number of files, adjust as needed
+    }
+});
 
 // ! Cars Add api
-router.post('/add', authCheck, async (req, res) => {  
+router.post('/add', upload.array('images'), apiAuthCheck, checkOrigin, async (req, res) => {  
+
+    try {
+        const carData = req.body; // Form data
+        const images = req.files; // Uploaded images
+
     
-    // if (res.locals.role != 'admin' && res.headersSent.apiKey != "admin") {
-    //     return res.status(400).send("Admin Role required to Access this route.")
-    // }
-    const dataArr = req.body;
-    dataArr.forEach(async (carData) => {
-        let {
-            name, 
-            brand, 
-            color, 
-            description, 
-            condition, 
-            doors, 
-            transmission, 
-            fuelType, 
-            seats, 
-            price, 
-            discount, 
-            status, 
-            imgUrls, 
-            createdTime, 
-            updateTime 
-        } = carData
+        // Upload images to S3
+        const uploadedImageUrls = await Promise.all(images.map(async (image) => {
+            const uploadParams = {
+                Bucket: 'reubz',
+                Key: `${carData.brand.toLowerCase()}-${carData.name.toLowerCase()}/${Math.floor(Math.random() * 100)}-${image.originalname.toLowerCase()}`, // You can set a different key as per your requirement
+                Body: image.buffer, // Image data
+                // ACL: 'public-read', // Make the uploaded image public
+                ContentType: image.mimetype // Mime type of the image
+            };
+    
+            const uploadResult = await s3.upload(uploadParams).promise();
+            return uploadResult.Location; // URL of the uploaded image
+        }));
+        // Add image URLs to carData
+        carData.imgUrls = uploadedImageUrls;
 
-        try {
-            const id = uuidv4()
-            await carsDB.create({
-                id: id,
-                name: name,
-                brand: brand,
-                color: color,
-                description: description,
-                condition: condition,
-                doors: doors,
-                transmission: transmission,
-                fuelType: fuelType, 
-                seats: seats,
-                price: price,
-                discount: discount,
-                status: status,
-                imgUrls: imgUrls,
-                createdTime: createdTime,
-                updateTime: updateTime,
-            }).then((car) => {
-                console.log(car.name + "saved ")
-            })
-        } catch (error) {
-            console.log('/api/cars.js - add car route - '+ error)
+        await carsDB.create(carData);
+        res.status(200).json({
+            error: { code: 100, message: "No errors" },
+            status: { code: 200, message: "Sucessfully Added New Car" }
+        })
+    } catch (error) {
+        console.log('/api/cars.js - add car route - '+ error)
 
-        }
-    })
-    res.status(200).send("Data Saved")
+    }
 })
 
 // ! Cars Delete API